# gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample API for demonstration.",
        "title": "Example API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/mri/devices": {
            "get": {
                "description": "получит список mri аппаратов",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mri"
                ],
                "summary": "получит список mri аппаратов",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "mri аппараты",
                        "schema": {
                            "$ref": "#/definitions/uzi.GetUziDeviceOut"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mri/echographics/{id}": {
            "get": {
                "description": "получает эхографику mri",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mri"
                ],
                "summary": "получает эхографику mri",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mri_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "echographics",
                        "schema": {
                            "$ref": "#/definitions/uzi.GetEchographicsOut"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Обновляет эхографику",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mri"
                ],
                "summary": "Обновляет эхографику",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mri_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "обновляемые значения",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/uzi.PatchEchographicsIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "echographic",
                        "schema": {
                            "$ref": "#/definitions/uzi.PatchEchographicsOut"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mri/images/{id}/nodes-segments": {
            "get": {
                "description": "получит ноды и сегменты на указанном изображении",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mri"
                ],
                "summary": "получит ноды и сегменты на указанном изображении",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "image_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "nodes\u0026\u0026segments",
                        "schema": {
                            "$ref": "#/definitions/uzi.GetUziNodeSegmentsOut"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mri/mris": {
            "post": {
                "description": "Загружает узи на обработку",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mri"
                ],
                "summary": "Загружает узи на обработку",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "mri file. (обязательно с .tiff/.png)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "проекция узи",
                        "name": "projection",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id пациента",
                        "name": "patient_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "id узи апапапапарата",
                        "name": "device_id",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "molodec",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mri/mris/{id}": {
            "get": {
                "description": "получает mri",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mri"
                ],
                "summary": "получает mri",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mri_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "mri",
                        "schema": {
                            "$ref": "#/definitions/internal_api_uzi.GetUziOut"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "patch": {
                "description": "Обновляет узи",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mri"
                ],
                "summary": "Обновляет узи",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mri_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "обновляемые значения",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/uzi.PatchUziIn"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "mri",
                        "schema": {
                            "$ref": "#/definitions/uzi.PatchUziOut"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mri/mris/{id}/images": {
            "get": {
                "description": "получает списк id кадров mri",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mri"
                ],
                "summary": "получает списк id кадров mri",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mri_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "images",
                        "schema": {
                            "$ref": "#/definitions/internal_api_uzi.GetUziImagesOut"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/mri/mris/{id}/nodes": {
            "get": {
                "description": "получить все узлы узи",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mri"
                ],
                "summary": "получить все узлы узи",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "mri_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "id узла",
                        "schema": {
                            "$ref": "#/definitions/internal_api_uzi.GetAllNodesOut"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/mri/patient/{id}/mris": {
            "get": {
                "description": "Получить узи пациента",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "mri"
                ],
                "summary": "Получить узи пациента",
                "parameters": [
                    {
                        "type": "string",
                        "description": "access_token",
                        "name": "token",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "patient_id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/uzi.GetPatientUziOut"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api_uzi.GetUziOut": {
            "type": "object",
            "properties": {
                "checked": {
                    "type": "boolean"
                },
                "create_at": {
                    "type": "string"
                },
                "device_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "patient_id": {
                    "type": "string"
                },
                "projection": {
                    "type": "string"
                }
            }
        },
...
        "uzi.GetUziDeviceOut": {
            "type": "object",
            "properties": {
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api_uzi.Device"
                    }
                }
            }
        },
        "uzi.GetUziNodeSegmentsOut": {
            "type": "object",
            "properties": {
                "nodes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api_uzi.Node"
                    }
                },
                "segments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api_uzi.Segment"
                    }
                }
            }
        },
...
    }
}